refactor(client): match export casing with export casing
